Make Azle compatible with CDK Framework fcee33f
