+++ src/compiler/typescript_to_rust/azle_generate/Cargo.toml
@@ -12,6 +12,6 @@ swc_ecma_parser = "0.117.4"
 derive-syn-parse = "0.1.5"
 swc_ecma_ast = "0.90.10"
 annotate-snippets = "0.9.1"
-cdk_framework = { git = "https://github.com/demergent-labs/cdk_framework", rev = "0ad949d010150f509e4395dcf377313be026eaf2" }
+cdk_framework = { git = "https://github.com/demergent-labs/cdk_framework", rev = "fcee33ffe2b66a3e744ba9a3c8e0f7fa4377c8ba" }
 # cdk_framework = { path = "../../../../../cdk_framework" } # For local rust-analyzer
 # cdk_framework = { path = "../../../../../../../cdk_framework" } # For local deployment in azle/examples
